\name{fasta.to.alignment}
\alias{fasta.to.alignment}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal function, used by sim.codon. }
\description{
   reads a text file containing sequences in the FASTA format, and returns
an object of class \code{"alignment"} from the \pkg{"ape"} package.
}
\usage{
fasta.to.alignment(fasta.file.name)
}
