\name{make.dat}
\alias{make.dat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal function, used by sim.codon. }
\description{
   Writes a file to be used by PAML to simulate codon sequences.
}
\usage{
make.dat(n.seq, n.codons, reps, tree, template.dat, file, freqs, omega, kap)
}
