\name{synDss-package}
\alias{synDss-package}
\alias{synDss}
\docType{package}
\title{synDss: Recombination detection with synonymous distances}
\description{
  \pkg{synDss} provides functions for performing recombination detection using synonymous distances
}
\author{
Peter Chi

Maintainer: Peter Chi <pchi@calpoly.edu>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }

